clean up App.tsx
