+++ src/Components/index.tsx
@@ -1,3 +1,4 @@
 export { default as Home } from './Home/Home';
 export { default as CreateVideo } from './CreateVideo/CreateVideo';
 export { Player } from './Import/Player';
+export { Modal } from './Modal/Modal';
